+++ src/net/sourceforge/fenixedu/domain/organizationalStructure/PersonFunction.java
@@ -6,7 +6,7 @@ import net.sourceforge.fenixedu.domain.exceptions.DomainException;
 
 public class PersonFunction extends PersonFunction_Base {
 
-    public void edit(IFunction function, Date beginDate, Date endDate, Integer credits) {
+    public void edit(IFunction function, Date beginDate, Date endDate, Double credits) {
         this.setFunction(function);
         this.setCredits(credits);        
         this.setBeginDate(beginDate);
